1 决定使用Stock.Indicators作为库 2 新增了相关测试用例 3 移除了Robot中的OHLCV相关类，维护工作量太大

+++ src/Quant.Infra.Net/SourceData/Model/Ohlcv.cs
@@ -1,6 +1,4 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Text;
 
 namespace Quant.Infra.Net.SourceData.Model
 {
@@ -16,7 +14,7 @@ namespace Quant.Infra.Net.SourceData.Model
 
         public decimal Close { get; set; }
 
-        public long Volume { get; set; }
+        public decimal Volume { get; set; }
 
         public decimal AdjustedClose { get; set; }
     }
